pyproject.toml
[tool.poetry]
name = "web-hw11"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "web_hw11"}]

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.10.4"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.12"
psycopg2 = "^2.9.6"
libgravatar = "^1.0.4"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
passlib = "^1.7.4"
cryptography = "^40.0.2"
bcrypt = "^4.0.1"
email-validator = "1.1"
fastapi-mail = "1.2.7"
fastapi-limiter = "^0.1.5"
aioredis = "^2.0.1"
cloudinary = "^1.33.0"
python-dotenv = "^1.0.0"
sphinx = "^7.0.1"
pytest-cov = "^4.1.0"
httpx = "^0.24.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

